fixup! Move ss, tf2ss, etc into statesp module
